pyproject.toml
[project]
name = "hdl_runner"
authors = [
    { name = "Alan Vekselman", email = "alanvek98@gmail.com" },
]
version = "0.0.1"
requires-python = ">=3.8"
dependencies = [
    "celosia @ git+https://github.com/AlanVek/celosia.git",
    "amaranth_yosys",
    "cocotb>=1.9.2,<=2.0.1",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.build]
includes = ["hdl_runner"]
